Fix item parsing of raw type 12
I had the problem that an eternal life potion could not be parsed
correctly. The reason for this was that `sub_ident` had the value
655360016. Since the value was greater than 16, it was incorrectly
assumed that this is a resource. However, it could not be parsed as a
resource either, since in this case the value can only be between 17 and
21. As a result, this particular eternal life potion was parsed as
`ItemType::Unknown`.

If you take a closer look at the number, it appears to be two separate
numbers: 65536 which is 2^16 and 0016. In fact, 16 is the correct value
for a eternal life potion, so 65536 must be ignored.

In binary notation, the number 655360016 looks like this:

  00100111 00010000 00000000 00010000
                             ^^^^^^^^

As one can see above, only the lower 8 bits are relevant. This commit
ensures that all other bits are ignored.