Implement trait `PartialEq` for `Command`
To do so it was necessary to implement this trait for the enums and
structs `Command` depends on as well.

This has been done to more effectively check for equality in unit tests.
The macro `matches!()` for example is not able to check if the fields of
the enum match as well.

Now the following is possible:

  let command = Command::FortressBuild {
      f_type: FortressBuildingType::Academy,
  };
  let building_type = FortressBuildingType::Academy;

  assert_eq!(
      command,
      Command::FortressBuild {
          f_type: building_type
      }
  );