Add function to check if a `DateTime` is in the future
The are a fair number of `DateTime`s in this API where one wants to know
if they are in the future or the past. This new function can be used to
determine this. To check if the given `datetime` is in the present or
the past, just negate the result.