Parse yesterday's daily reward of Hellevator
The logic to parse yesterday's daily rewards was already implemented but
commented out. It is not clear to me why this has been done, it seems to
work fine.

The format of yesterday's daily rewards differs from today's, it has a
length of 14 instead of 10. The condition has been adjusted to handle
both formats respectively.