Uncomment method active_mut() of `HellevatorEvent`
This method can be used to obtain a mutable reference to a `Hellevator`
instance.

The game state has to be updated manually, after claiming yesterday's
daily rewards. This method can be used for this.

Why this code was commented out is not clear to me, it seems to work
fine.